remove context.CheckHeader, epxort server.handler* to file
